Avoid problems with Julia HTTP/SSL libs and fixup report
Switch to accessing the OpenQA API through the Perl client script. This avoids
having to fix at least one bug in EmbedTLS and also avoids reimplementing
OpenQA's authentication method which I know to be a pain.

I think that in principal the underlying problems in Julia HTTP and OpenQA
should be addressed. On the other hand using the official client script may be
preferable anyway to avoid creating Julia client libraries for many different
trackers. At least at this stage in the project.