Convert diff report to use BuildMatrix and BuildMatrixGrouped
Results are now less filtered, but similar rows are collapsed/elided. It is
now also possible to click on a test result to go through to OpenQA and bad
results are highlighted.

Notifications are now less verbose and instead rely on linking through to the
correct section of the status diff report. Notifications will not be sent at
all for test categories which have a high level of churn because they are not
effective when sent constantly.