Add lazy loading or linking interface and use it for OpenQA results
Currently JobResults don't contain any job group info and job groups don't
store which instance they are from. This introduces a generic interface for
specifying relations between structures in the database where it is not
appropriate to simply embed one structure within another.

The interface doubles as a form of lazy loading which can be used for other
purposes.