Optimise heavily used mload function
Inline the parsing of the Redis response to ensure 'type stability' and to
avoid, large, unnecessary allocations. Implementing this upstream, in the
Redis.jl package, would probably require a large rewrite and changes to the
API.

It is unlikely that we will need to do this for any other Redis calls.