Introduce quote syntax for bug tags and use it for propagated tags
It is nice to wrap test names with `backticks` because sometimes the names
contain an underscore which MD interprets as italics. However this broke the
bug tag parser.

This patch allows the parser to omit the backticks from the test name
token. However it marks it as 'quoted' and this is then used to indicate that
the bug tag was propagated.

Propagated tags are ignored during future propagations so that the original
tag is linked in the propagation.