Introduce Metarules for specifying arbitrary bug propagation rules
This allows one to write something like `[flavor=GCE]` in a issue or bug
title. This will prevent the bug from being propagated to OpenQA jobs which
do not have `FLAVOR` set to `GCE`. One can also write `[product~"^sle-12"]`
which will check if the regex "^sle-12" occurs in `product`.